Cache empty node in cached node manager

+++ rust/src/node_manager/cached_node_manager.rs
@@ -12,7 +12,7 @@ use std::{
     cmp::Eq,
     hash::Hash,
     ops::{Deref, DerefMut},
-    sync::{Arc, RwLockReadGuard, RwLockWriteGuard},
+    sync::{Arc, RwLock, RwLockReadGuard, RwLockWriteGuard},
 };
 
 use crate::{
@@ -22,6 +22,7 @@ use crate::{
         lock_cache::{EvictionHooks, LockCache},
     },
     storage::{Checkpointable, Storage},
+    types::{HasEmptyId, HasEmptyNode},
 };
 
 /// A wrapper which dereferences to [`N`] and additionally stores its dirty status,
@@ -86,21 +87,24 @@ impl<S: Storage> Deref for StorageEvictionHandler<S> {
 ///
 /// Nodes are retrieved from the underlying storage if they are not present in the cache, and stored
 /// back when they get evicted and they have been modified.
+#[cfg_attr(not(test), expect(unused))]
 pub struct CachedNodeManager<S>
 where
     S: Storage,
+    S::Item: HasEmptyNode,
 {
     // Cache for storing nodes in memory.
     nodes: LockCache<S::Id, NodeWithMetadata<S::Item>>,
     // Storage for managing IDs, fetching missing nodes, and storing evicted nodes.
     storage: Arc<StorageEvictionHandler<S>>,
+    empty_node: RwLock<NodeWithMetadata<S::Item>>,
 }
 
 impl<S> CachedNodeManager<S>
 where
     S: Storage + 'static,
     S::Id: Eq + Hash + Copy,
-    S::Item: Default,
+    S::Item: Default + HasEmptyNode,
 {
     /// Creates a new [`CachedNodeManager`] with the given capacity, storage backend, and pin
     /// predicate.
@@ -119,6 +123,10 @@ where
                     as Arc<dyn EvictionHooks<Key = S::Id, Value = NodeWithMetadata<S::Item>>>,
             ),
             storage,
+            empty_node: RwLock::new(NodeWithMetadata {
+                node: S::Item::make_empty_node(),
+                is_dirty: false,
+            }),
         }
     }
 }
@@ -126,13 +134,17 @@ where
 impl<S> NodeManager for CachedNodeManager<S>
 where
     S: Storage + 'static,
-    S::Id: Eq + Hash + Copy,
-    S::Item: Default,
+    S::Id: Eq + Hash + Copy + HasEmptyId,
+    S::Item: Default + HasEmptyNode,
 {
     type Id = S::Id;
     type NodeType = S::Item;
 
     fn add(&self, node: Self::NodeType) -> Result<Self::Id, Error> {
+        if node.is_empty_node() {
+            // No need to cache empty nodes
+            return Ok(Self::Id::make_empty_id());
+        }
         let id = self.storage.reserve(&node);
         let _guard = self.nodes.get_read_access_or_insert(id, move || {
             Ok(NodeWithMetadata {
@@ -150,6 +162,10 @@ where
         &self,
         id: Self::Id,
     ) -> Result<RwLockReadGuard<'_, impl Deref<Target = Self::NodeType>>, Error> {
+        if id.is_empty_id() {
+            return Ok(self.empty_node.read().unwrap());
+        }
+
         let lock = self.nodes.get_read_access_or_insert(id, || {
             let node = self.storage.storage.get(id)?;
             Ok(NodeWithMetadata {
@@ -167,6 +183,10 @@ where
         &self,
         id: Self::Id,
     ) -> Result<RwLockWriteGuard<'_, impl DerefMut<Target = Self::NodeType>>, Error> {
+        if id.is_empty_id() {
+            return Ok(self.empty_node.write().unwrap());
+        }
+
         let lock = self.nodes.get_write_access_or_insert(id, || {
             let node = self.storage.storage.get(id)?;
             Ok(NodeWithMetadata {
@@ -183,6 +203,10 @@ where
     /// It is not safe to call this function multiple times for the same ID, unless allowed by
     /// [`Self::S`].
     fn delete(&self, id: Self::Id) -> Result<(), Error> {
+        if id.is_empty_id() {
+            return Ok(());
+        }
+
         self.nodes.remove(id)?;
         self.storage.delete(id)?;
         Ok(())
@@ -193,7 +217,7 @@ impl<S> Checkpointable for CachedNodeManager<S>
 where
     S: Storage + 'static + Checkpointable,
     S::Id: Eq + Hash + Copy + Send + Sync,
-    S::Item: Default + Clone + Send + Sync,
+    S::Item: Default + Clone + Send + Sync + HasEmptyNode,
 {
     fn checkpoint(&self) -> Result<(), crate::storage::Error> {
         for (id, mut guard) in self.nodes.iter_write() {
@@ -221,6 +245,32 @@ mod tests {
 
     type TestNodeId = u32;
     type TestNode = i32;
+
+    impl HasEmptyNode for TestNode {
+        fn is_empty_node(&self) -> bool {
+            *self == i32::MIN
+        }
+
+        fn make_empty_node() -> Self
+        where
+            Self: Sized,
+        {
+            i32::MIN
+        }
+    }
+
+    impl HasEmptyId for TestNodeId {
+        fn is_empty_id(&self) -> bool {
+            *self == u32::MIN
+        }
+
+        fn make_empty_id() -> Self
+        where
+            Self: Sized,
+        {
+            u32::MIN
+        }
+    }
 
     /// Helper function to return a [`storage::Error::NotFound`] wrapped in an [`Error`]
     fn not_found() -> Result<NodeWithMetadata<TestNode>, Error> {
@@ -258,6 +308,17 @@ mod tests {
         assert_eq!(node_res.node, node);
     }
 
+    #[test]
+    fn cached_node_manager_add_returns_empty_id_for_empty_node() {
+        let node = TestNode::make_empty_node();
+        let mut storage = MockCachedNodeManagerStorage::new();
+        storage.expect_reserve().never(); // Shouldn't reserve ID for empty node
+        storage.expect_get().never(); // Shouldn't query storage on add
+        let manager = CachedNodeManager::new(10, storage, pin_nothing);
+        let id = manager.add(node).unwrap();
+        assert_eq!(id, TestNodeId::make_empty_id());
+    }
+
     #[rstest_reuse::apply(get_method)]
     fn cached_node_manager_get_methods_return_cached_entry(#[case] get_method: GetMethod) {
         let id = 0;
@@ -287,6 +348,17 @@ mod tests {
         let manager = CachedNodeManager::new(10, storage, pin_nothing);
         let entry = get_method(&manager, id).unwrap();
         assert!(entry == expected_entry);
+    }
+
+    #[rstest_reuse::apply(get_method)]
+    fn cached_node_manager_get_methods_return_empty_node_for_empty_id(
+        #[case] get_method: GetMethod,
+    ) {
+        let mut storage = MockCachedNodeManagerStorage::new();
+        storage.expect_get().never(); // Shouldn't query storage for empty ID
+        let manager = CachedNodeManager::new(10, storage, pin_nothing);
+        let entry = get_method(&manager, TestNodeId::make_empty_id()).unwrap();
+        assert!(entry == TestNode::make_empty_node());
     }
 
     #[rstest_reuse::apply(get_method)]
@@ -375,6 +447,16 @@ mod tests {
         assert_eq!(manager.nodes.iter_write().count(), 1);
         manager.delete(id).unwrap();
         assert_eq!(manager.nodes.iter_write().count(), 0);
+    }
+
+    #[test]
+    fn cached_node_manager_delete_on_empty_id_is_noop() {
+        let mut storage = MockCachedNodeManagerStorage::new();
+        storage.expect_delete().never();
+
+        let manager = CachedNodeManager::new(2, storage, pin_nothing);
+        // Shouldn't error
+        manager.delete(TestNodeId::make_empty_id()).unwrap();
     }
 
     #[test]
